feat: 비지니스 로직 ver.2

+++ src/main/java/com/example/demo/api/favoriteStock/service/FavoriteStockUseCase.java
@@ -2,8 +2,10 @@ package com.example.demo.api.favoriteStock.service;
 
 import com.example.demo.common.annotation.UseCase;
 import com.example.demo.domain.favoriteStock.entity.FavoriteStock;
+import com.example.demo.domain.favoriteStock.exception.FavoriteStockHandler;
 import com.example.demo.domain.favoriteStock.service.FavoriteStockCommandService;
 import com.example.demo.domain.favoriteStock.service.FavoriteStockQueryService;
+import com.example.demo.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -24,7 +26,27 @@ public class FavoriteStockUseCase {
         return favoriteStockQueryService.getUserFavoriteStockListByPage(userId, pageable);
     }
 
-    public Void deleteFavoriteStock(Long favoriteStockId) {
-        return favoriteStockCommandService.deleteFavoriteStock(favoriteStockId);
+    public Long updateFavoriteStockAlarm(User user, Long favoriteStockId, boolean isEnabled) {
+        FavoriteStock favoriteStock = favoriteStockQueryService.getFavoriteStockById(favoriteStockId);
+
+        if (favoriteStock.getUser().getId().equals(user.getId())) {
+            favoriteStockCommandService.updateFavoriteStockAlarm(favoriteStockId, isEnabled);
+
+            return favoriteStockId;
+        }
+        else {
+            throw FavoriteStockHandler.FORBIDDEN;
+        }
+    }
+
+    public void deleteFavoriteStock(User user, Long favoriteStockId) {
+        FavoriteStock favoriteStock = favoriteStockQueryService.getFavoriteStockById(favoriteStockId);
+
+        if (favoriteStock.getUser().getId().equals(user.getId())) {
+            favoriteStockCommandService.deleteFavoriteStock(favoriteStockId);
+        }
+        else {
+            throw FavoriteStockHandler.FORBIDDEN;
+        }
     }
 }
